+++ QoolUI/QoolControls/Components/BasicText.qml
@@ -1,0 +1,8 @@
+import QtQuick
+import Qool
+
+Text {
+    id: root
+    color: Style.text
+    font.pixelSize: Style.textSize
+}